+++ src/main/java/io/dataspaceconnector/services/messages/types/ContractRequestService.java
@@ -15,15 +15,13 @@
  */
 package io.dataspaceconnector.services.messages.types;
 
-import java.net.URI;
-import java.util.Map;
-
 import de.fraunhofer.iais.eis.ContractAgreementMessageImpl;
 import de.fraunhofer.iais.eis.ContractRequest;
 import de.fraunhofer.iais.eis.ContractRequestMessageBuilder;
 import de.fraunhofer.iais.eis.Message;
 import de.fraunhofer.iais.eis.util.ConstraintViolationException;
 import de.fraunhofer.iais.eis.util.Util;
+import de.fraunhofer.ids.messaging.util.IdsMessageUtils;
 import io.dataspaceconnector.exceptions.MessageException;
 import io.dataspaceconnector.exceptions.MessageResponseException;
 import io.dataspaceconnector.exceptions.RdfBuilderException;
@@ -33,7 +31,8 @@ import io.dataspaceconnector.utils.IdsUtils;
 import io.dataspaceconnector.utils.Utils;
 import org.springframework.stereotype.Service;
 
-import static de.fraunhofer.isst.ids.framework.util.IDSUtils.getGregorianNow;
+import java.net.URI;
+import java.util.Map;
 
 /**
  * Message service for ids contract request messages.
@@ -43,7 +42,9 @@ public final class ContractRequestService
         extends AbstractMessageService<ContractRequestMessageDesc> {
 
     /**
-     * @throws IllegalArgumentException If desc is null.
+     * @throws IllegalArgumentException     if desc is null.
+     * @throws ConstraintViolationException if security tokes is null or another error appears
+     *                                      when building the message.
      */
     @Override
     public Message buildMessage(final ContractRequestMessageDesc desc)
@@ -58,7 +59,7 @@ public final class ContractRequestService
         final var contractId = desc.getTransferContract();
 
         return new ContractRequestMessageBuilder()
-                ._issued_(getGregorianNow())
+                ._issued_(IdsMessageUtils.getGregorianNow())
                 ._modelVersion_(modelVersion)
                 ._issuerConnector_(connectorId)
                 ._senderAgent_(connectorId)
@@ -79,9 +80,9 @@ public final class ContractRequestService
      * @param recipient The recipient.
      * @param request   The contract request.
      * @return The response map.
-     * @throws MessageException         If message handling failed.
-     * @throws RdfBuilderException      If the contract request rdf string could not be built.
-     * @throws IllegalArgumentException If contract request is null.
+     * @throws MessageException         if message handling failed.
+     * @throws RdfBuilderException      if the contract request rdf string could not be built.
+     * @throws IllegalArgumentException if contract request is null.
      */
     public Map<String, String> sendMessage(final URI recipient, final ContractRequest request)
             throws MessageException, RdfBuilderException {