+++ src/main/java/io/dataspaceconnector/utils/MappingUtils.java
@@ -199,7 +199,7 @@ public final class MappingUtils {
 
         if (description != null) {
             desc.setDescription(description.size() == 1 ? description.get(0).toString()
-                                                        : description.toString());
+                    : description.toString());
         }
 
         if (title != null) {
@@ -255,9 +255,9 @@ public final class MappingUtils {
      * Adds the string value of a given list as an additional property. If the list only contains
      * one element, the string value will not contain brackets.
      *
-     * @param list the list.
+     * @param list       the list.
      * @param additional the map of additional properties.
-     * @param key the map key to use.
+     * @param key        the map key to use.
      */
     private static void addListToAdditional(final List<?> list,
                                             final Map<String, String> additional,
@@ -423,9 +423,9 @@ public final class MappingUtils {
      *
      * @param rule The ids rule.
      * @return The rule template.
-     * @throws IllegalArgumentException if the rule is null.
-     * @throws io.dataspaceconnector.exceptions.RdfBuilderException
-     *         if the rule cannot be converted to string.
+     * @throws IllegalArgumentException                             if the rule is null.
+     * @throws io.dataspaceconnector.exceptions.RdfBuilderException if the rule cannot be
+     * converted to string.
      */
     public static RuleTemplate fromIdsRule(final Rule rule) {
         Utils.requireNonNull(rule, ErrorMessages.ENTITY_NULL);