+++ src/test/java/io/dataspaceconnector/controller/messages/ConnectorUpdateMessageControllerTest.java
@@ -15,14 +15,14 @@
  */
 package io.dataspaceconnector.controller.messages;
 
-import de.fraunhofer.isst.ids.framework.communication.broker.IDSBrokerService;
-import de.fraunhofer.isst.ids.framework.configuration.ConfigurationUpdateException;
+import de.fraunhofer.iais.eis.DynamicAttributeToken;
+import de.fraunhofer.iais.eis.DynamicAttributeTokenBuilder;
+import de.fraunhofer.iais.eis.MessageProcessedNotificationMessageBuilder;
+import de.fraunhofer.iais.eis.TokenFormat;
+import de.fraunhofer.ids.messaging.broker.IDSBrokerService;
+import de.fraunhofer.ids.messaging.core.config.ConfigUpdateException;
+import de.fraunhofer.ids.messaging.protocol.multipart.mapping.MessageProcessedNotificationMAP;
 import io.dataspaceconnector.services.ids.ConnectorService;
-import okhttp3.MediaType;
-import okhttp3.Protocol;
-import okhttp3.Request;
-import okhttp3.Response;
-import okhttp3.ResponseBody;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,7 +32,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
+import javax.xml.datatype.DatatypeFactory;
 import java.io.IOException;
+import java.net.URI;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -53,6 +55,10 @@ public class ConnectorUpdateMessageControllerTest {
     private MockMvc mockMvc;
 
     private final String recipient = "https://someURL";
+    private final DynamicAttributeToken token = new DynamicAttributeTokenBuilder()
+            ._tokenValue_("token")
+            ._tokenFormat_(TokenFormat.JWT)
+            .build();
 
     @Test
     public void sendConnectorUpdateMessage_unauthorized_returnUnauthorized() throws Exception {
@@ -61,90 +67,90 @@ public class ConnectorUpdateMessageControllerTest {
 
     @Test
     @WithMockUser("ADMIN")
-    public void sendConnectorUpdateMessage_noRecipient_throws400()
-            throws Exception {
+    public void sendConnectorUpdateMessage_noRecipient_throws400() throws Exception {
         /* ARRANGE */
         // Nothing to arrange here.
 
         /* ACT */
-        final var result = mockMvc.perform(post("/api/ids/connector/update")).andExpect(status().isBadRequest()).andReturn();
+        final var result = mockMvc.perform(post("/api/ids/connector/update"))
+                .andReturn();
 
         /* ASSERT */
         assertTrue(result.getResponse().getContentAsString().isEmpty());
+        assertEquals(400, result.getResponse().getStatus());
     }
 
     @Test
     @WithMockUser("ADMIN")
-    public void sendConnectorUpdateMessage_failUpdateConfigModel_throws500()
-            throws Exception {
+    public void sendConnectorUpdateMessage_failUpdateConfigModel_throws500() throws Exception {
         /* ARRANGE */
-        Mockito.doThrow(ConfigurationUpdateException.class).when(connectorService).updateConfigModel();
+        Mockito.doReturn(token).when(connectorService).getCurrentDat();
+        Mockito.doThrow(ConfigUpdateException.class).when(connectorService).updateConfigModel();
 
         /* ACT */
         final var result = mockMvc.perform(post("/api/ids/connector/update")
-                                                   .param("recipient", recipient))
-                                  .andExpect(status().isInternalServerError()).andReturn();
+                .param("recipient", recipient)).andReturn();
 
         /* ASSERT */
         assertEquals("Failed to update configuration.", result.getResponse().getContentAsString());
+        assertEquals(500, result.getResponse().getStatus());
     }
 
     @Test
     @WithMockUser("ADMIN")
-    public void sendConnectorUpdateMessage_failUpdateAtBroker_throws500()
-            throws Exception {
+    public void sendConnectorUpdateMessage_failUpdateAtBroker_throws500() throws Exception {
         /* ARRANGE */
-        Mockito.doThrow(IOException.class).when(brokerService).updateSelfDescriptionAtBroker(Mockito.eq(recipient));
+        Mockito.doReturn(token).when(connectorService).getCurrentDat();
+        Mockito.doThrow(IOException.class).when(brokerService).updateSelfDescriptionAtBroker(Mockito.eq(URI.create(recipient)));
 
         /* ACT */
         final var result = mockMvc.perform(post("/api/ids/connector/update")
-                                                   .param("recipient", recipient))
-                                  .andExpect(status().isInternalServerError()).andReturn();
+                .param("recipient", recipient)).andReturn();
 
         /* ASSERT */
-        assertEquals("Ids message handling failed. null", result.getResponse().getContentAsString());
+        assertEquals(500, result.getResponse().getStatus());
     }
 
     @Test
     @WithMockUser("ADMIN")
-    public void sendConnectorUpdateMessage_brokerEmptyResponseBody_throws500()
-            throws Exception {
+    public void sendConnectorUpdateMessage_brokerEmptyResponseBody_throws500() throws Exception {
         /* ARRANGE */
-        final var response =
-                new Response.Builder().request(new Request.Builder().url(recipient).build())
-                                      .protocol(Protocol.HTTP_1_1).code(200).message("").build();
-        Mockito.doReturn(response).when(brokerService).updateSelfDescriptionAtBroker(Mockito.eq(recipient));
+        Mockito.doReturn(token).when(connectorService).getCurrentDat();
+        Mockito.doThrow(IOException.class).when(brokerService).updateSelfDescriptionAtBroker(Mockito.eq(URI.create(recipient)));
 
         /* ACT */
         final var result = mockMvc.perform(post("/api/ids/connector/update")
-                                                   .param("recipient", recipient))
-                                  .andExpect(status().isInternalServerError()).andReturn();
+                .param("recipient", recipient)).andReturn();
 
         /* ASSERT */
-        assertEquals("Ids message handling failed. null", result.getResponse().getContentAsString());
+        assertEquals(500, result.getResponse().getStatus());
     }
 
     @Test
     @WithMockUser("ADMIN")
-    public void sendConnectorUpdateMessage_validRequest_returnsBrokerResponse()
-            throws Exception {
+    public void sendConnectorUpdateMessage_validRequest_returnsBrokerResponse() throws Exception {
         /* ARRANGE */
-        final var response =
-                new Response.Builder().request(new Request.Builder().url(recipient).build())
-                                      .protocol(
-                                              Protocol.HTTP_1_1).code(200).message("")
-                                      .body(ResponseBody.create("ANSWER", MediaType
-                                              .parse("application/text")))
-                                      .build();
+        final var message = new MessageProcessedNotificationMessageBuilder()
+                ._issuerConnector_(new URI("https://url"))
+                ._correlationMessage_(new URI("https://cormessage"))
+                ._issued_(DatatypeFactory.newInstance()
+                        .newXMLGregorianCalendar("2009-05-07T17:05:45.678Z"))
+                ._senderAgent_(new URI("https://sender"))
+                ._modelVersion_("4.0.0")
+                ._securityToken_(token)
+                .build();
 
-        Mockito.doReturn(response).when(brokerService).updateSelfDescriptionAtBroker(Mockito.eq(recipient));
+        final var response = new MessageProcessedNotificationMAP(message);
+
+        Mockito.doReturn(token).when(connectorService).getCurrentDat();
+        Mockito.doReturn(response).when(brokerService).updateSelfDescriptionAtBroker(Mockito.eq(URI.create(recipient)));
 
         /* ACT */
         final var result = mockMvc.perform(post("/api/ids/connector/update")
-                                                   .param("recipient", recipient))
-                                  .andExpect(status().isOk()).andReturn();
+                .param("recipient", recipient)).andReturn();
 
         /* ASSERT */
-        assertEquals("ANSWER", result.getResponse().getContentAsString());
+        assertEquals("", result.getResponse().getContentAsString());
+        assertEquals(200, result.getResponse().getStatus());
     }
 }