+++ src/test/java/io/dataspaceconnector/services/EntityUpdateServiceTest.java
@@ -15,14 +15,6 @@
  */
 package io.dataspaceconnector.services;
 
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-
 import de.fraunhofer.iais.eis.Artifact;
 import de.fraunhofer.iais.eis.ArtifactBuilder;
 import de.fraunhofer.iais.eis.Representation;
@@ -47,6 +39,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -59,7 +59,10 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest(classes = {EntityUpdateService.class, RequestedResourceFactory.class})
+@SpringBootTest(classes = {
+        EntityUpdateService.class,
+        RequestedResourceFactory.class
+})
 public class EntityUpdateServiceTest {
 
     @MockBean
@@ -247,9 +250,9 @@ public class EntityUpdateServiceTest {
                 .add(eq(agreementId), eq(Set.of(artifactId1, artifactId2)));
     }
 
-    /**************************************************************************
-     * Utilities.
-     *************************************************************************/
+    /***********************************************************************************************
+     * Utilities.                                                                                  *
+     **********************************************************************************************/
 
     private Resource getResource() {
         return new ResourceBuilder(URI.create("https://resource-id.com"))
@@ -271,5 +274,4 @@ public class EntityUpdateServiceTest {
     private RequestedResource getRequestedResource() {
         return requestedResourceFactory.create(new RequestedResourceDesc());
     }
-
 }