+++ src/test/java/io/dataspaceconnector/services/ids/IdsDutyBuilderTest.java
@@ -29,7 +29,7 @@ import io.dataspaceconnector.model.rule.ContractRule;
 import io.dataspaceconnector.model.rule.ContractRuleDesc;
 import io.dataspaceconnector.model.rule.ContractRuleFactory;
 import io.dataspaceconnector.services.ids.builder.IdsDutyBuilder;
-import de.fraunhofer.isst.ids.framework.configuration.SerializerProvider;
+import de.fraunhofer.ids.messaging.util.SerializerProvider;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -190,7 +190,8 @@ public class IdsDutyBuilderTest {
                 + "    } ],\n"
                 + "      \"ids:constraint\" : [ {\n"
                 + "        \"@type\" : \"ids:Constraint\",\n"
-                + "        \"@id\" : \"https://w3id.org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
+                + "        \"@id\" : \"https://w3id" +
+                ".org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
                 + "-499eec24bc57\",\n"
                 + "      \"ids:rightOperand\" : {\n"
                 + "        \"@value\" : \"2020-07-11T00:00:00Z\",\n"
@@ -218,7 +219,8 @@ public class IdsDutyBuilderTest {
                 + "    } ],\n"
                 + "      \"ids:constraint\" : [ {\n"
                 + "        \"@type\" : \"ids:Constraint\",\n"
-                + "        \"@id\" : \"https://w3id.org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
+                + "        \"@id\" : \"https://w3id" +
+                ".org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
                 + "-499eec24bc57\",\n"
                 + "      \"ids:rightOperand\" : {\n"
                 + "        \"@value\" : \"2020-07-11T00:00:00Z\",\n"
@@ -248,7 +250,8 @@ public class IdsDutyBuilderTest {
                 + "    } ],\n"
                 + "      \"ids:constraint\" : [ {\n"
                 + "        \"@type\" : \"ids:Constraint\",\n"
-                + "        \"@id\" : \"https://w3id.org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
+                + "        \"@id\" : \"https://w3id" +
+                ".org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
                 + "-499eec24bc57\",\n"
                 + "      \"ids:rightOperand\" : {\n"
                 + "        \"@value\" : \"2020-07-11T00:00:00Z\",\n"
@@ -275,7 +278,8 @@ public class IdsDutyBuilderTest {
                 + "-aa4c29d29fb5\",\n"
                 + "    \"ids:constraint\" : [ {\n"
                 + "      \"@type\" : \"ids:Constraint\",\n"
-                + "      \"@id\" : \"https://w3id.org/idsa/autogen/constraint/f2acf67f-bc4c-4e64-87fc"
+                + "      \"@id\" : \"https://w3id.org/idsa/autogen/constraint/f2acf67f-bc4c-4e64" +
+                "-87fc"
                 + "-499eec24bc57\",\n"
                 + "      \"ids:rightOperand\" : {\n"
                 + "        \"@value\" : \"2020-07-11T00:00:00Z\",\n"