+++ CHANGELOG.md
@@ -1,15 +1,32 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [5.2.1] - 2021-07-02
+## [6.0.0] - XXXX-XX-XX
 
 ### Added
-- Make bootstrapping feature optional. It can be disabled in the `application.properties`.
+- Provide REST endpoint for full-text search at the IDS Broker: `/ids/search`.
+- Check if the issuer connector of an artifact request does correspond to the signed consumer of the
+  transfer contract.
+
+### Changed
+- Replace IDS Connector Framework v5.0.4 by IDS Messaging Services v4.1.1.2.
+- Edit response codes and response content for the following endpoints: `/ids/connector/unavailable`,
+  `/ids/connector/update`, `/ids/resource/unavailable`, `/ids/resource/update`, `/ids/query`.
+- Move implementation for sending IDS query, connector, and resource messages to
+  `GeneralMessageService`.
+- Handle DAT retrieving errors in `PRODUCTIVE_DEPLOYMENT` with status code 500 and a corresponding
+  message.
+- Change naming of the resource's license attribute from `licence` to `license`.
+
+## [5.2.1] - 2021-07-02
 
 ### Changed
 - Increase spring-boot version to 2.5.2.
 - Increase checkstyle version to 8.44.
 - Increase pmd version to 6.36.0.
+
+### Fixed
+- Make bootstrapping feature optional. It can be enabled in the `application.properties`.
 
 ## [5.2.0] - 2021-06-23